+++ src/card/cards/vintage_cube.rs
@@ -77,3 +77,4 @@ pub const UNHOLY_HEAT: CardDefinitionId = CardDefinitionId(2159);
 pub const JACE_WIELDER_OF_MYSTERIES: CardDefinitionId = CardDefinitionId(2160);
 pub const DEEP_CAVERN_BAT: CardDefinitionId = CardDefinitionId(2161);
 pub const VIVI_ORNITIER: CardDefinitionId = CardDefinitionId(2162);
+pub const FLAME_OF_ANOR: CardDefinitionId = CardDefinitionId(2163);